backend/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
pyyaml>=6.0
dataclasses>=0.6

# 主题建模和向量计算依赖
gensim>=4.3.0
faiss-cpu>=1.7.4
numpy>=1.21.0
scikit-learn>=1.0.0
jieba>=0.42.1
nltk>=3.8

# FastAPI服务器依赖
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# 数据库和认证依赖
supabase>=2.0.0
pyjwt>=2.8.0

# 其他工具依赖
python-json-logger>=2.0.0